Add access control to domains

+++ src/main/java/com/arnellconsulting/worktajm/web/rest/DomainResource.java
@@ -1,5 +1,6 @@
 package com.arnellconsulting.worktajm.web.rest;
 
+import com.arnellconsulting.worktajm.domain.User;
 import com.arnellconsulting.worktajm.security.AuthoritiesConstants;
 import com.arnellconsulting.worktajm.service.UserService;
 import com.codahale.metrics.annotation.Timed;
@@ -20,6 +21,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
@@ -30,6 +32,8 @@ import java.net.URISyntaxException;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import static org.elasticsearch.index.query.QueryBuilders.*;
 
@@ -97,9 +101,11 @@ public class DomainResource {
     @Secured(AuthoritiesConstants.ADMIN)
     public ResponseEntity<DomainDTO> updateDomain(@Valid @RequestBody DomainDTO domainDTO) throws URISyntaxException {
         log.debug("REST request to update Domain : {}", domainDTO);
-        if (domainDTO.getId() == null) {
-            return createDomain(domainDTO);
-        }
+
+        // Check user has permission to access domain
+        checkAccessToDomain(domainDTO.getId());
+
+        // Update
         Domain domain = domainMapper.toEntity(domainDTO);
         domain = domainRepository.save(domain);
         DomainDTO result = domainMapper.toDto(domain);
@@ -120,7 +126,8 @@ public class DomainResource {
     @Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.USER})
     public ResponseEntity<List<DomainDTO>> getAllDomains(Pageable pageable) {
         log.debug("REST request to get a page of Domains");
-        Page<Domain> page = domainRepository.findAll(pageable);
+        Optional<User> user = this.userService.getUserWithAuthorities();
+        Page<Domain> page = domainRepository.findAllByAuthorizedUsersIs(user.get(), pageable);
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/domains");
         return new ResponseEntity<>(domainMapper.toDto(page.getContent()), headers, HttpStatus.OK);
     }
@@ -136,6 +143,10 @@ public class DomainResource {
     @Secured({AuthoritiesConstants.ADMIN, AuthoritiesConstants.USER})
     public ResponseEntity<DomainDTO> getDomain(@PathVariable Long id) {
         log.debug("REST request to get Domain : {}", id);
+
+        // Check user has permission to access domain
+        checkAccessToDomain(id);
+
         Domain domain = domainRepository.findOneWithEagerRelationships(id);
         DomainDTO domainDTO = domainMapper.toDto(domain);
         return ResponseUtil.wrapOrNotFound(Optional.ofNullable(domainDTO));
@@ -152,6 +163,10 @@ public class DomainResource {
     @Secured(AuthoritiesConstants.ADMIN)
     public ResponseEntity<Void> deleteDomain(@PathVariable Long id) {
         log.debug("REST request to delete Domain : {}", id);
+
+        // Check user has permission to access domain
+        checkAccessToDomain(id);
+
         domainRepository.delete(id);
         domainSearchRepository.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
@@ -174,4 +189,23 @@ public class DomainResource {
         return new ResponseEntity<>(domainMapper.toDto(page.getContent()), headers, HttpStatus.OK);
     }
 
+    private void checkAccessToDomain(Long domainId) {
+
+        // Must have id
+        if (domainId == null) {
+            throw new AccessDeniedException("No id");
+        }
+
+        // Must find user
+        Optional<User> user = this.userService.getUserWithAuthorities();
+        if (!user.isPresent()) {
+            throw new AccessDeniedException("User not found");
+        }
+
+        // Make sure user has access to domain
+        Domain existingDomain = domainRepository.findOne(domainId);
+        if (!existingDomain.isUserIdAuthorized(user.get().getId())) {
+            throw new AccessDeniedException("User is not allowed to access domain");
+        }
+    }
 }
